// SDKs/twilio.com/twilio_microvisor_v1/1.36.0/typescript/src/sdk/models/shared/microvisorv1app.ts
import { Metadata, SpeakeasyBase } from "../../../internal/utils/utils";


export class MicrovisorV1App extends SpeakeasyBase {
  @Metadata({ data: "json, name=account_sid" })
  accountSid?: string;

  @Metadata({ data: "json, name=date_created" })
  dateCreated?: Date;

  @Metadata({ data: "json, name=date_updated" })
  dateUpdated?: Date;

  @Metadata({ data: "json, name=hash" })
  hash?: string;

  @Metadata({ data: "json, name=sid" })
  sid?: string;

  @Metadata({ data: "json, name=unique_name" })
  uniqueName?: string;

  @Metadata({ data: "json, name=url" })
  url?: string;
}